move some modules into `qulab`
