+++ qulab/__init__.py
@@ -1,0 +1,1 @@
+from ._version import __version__
